moved schema setup to separate files
