+++ lib/schema-add-hooks.js
@@ -1,0 +1,9 @@
+module.exports = function (schema, options) {
+  schema.pre('save', function (next) {
+    if (options.usernameLowerCase && this[options.usernameField]) {
+      this[options.usernameField] = this[options.usernameField].toLowerCase();
+    }
+
+    next();
+  });
+};
